POC-Compensar-App-Dev/search-website/api/requirements.txt
azure-core==1.13.0
azure-functions==1.7.0
azure-search-documents==11.1.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
typing-extensions==3.10.0.0
urllib3==1.26.5
